autoUpdateImage/requirements.txt
qingcloud-sdk
